app/requirements.txt
Django>=5.0,<5.1
gunicorn>=22.0
whitenoise>=6.6
python-dotenv>=1.0
Pillow>=10.0
google-cloud-aiplatform
google-cloud-storage
